Dropped and re-created database model classes based on "Phase 2" database. Any Data Annotations created for these models have been lost.
VendorController modified to support BrowseVendors View.

+++ cs3750Fall2013/RecreationOutlet/Warehouse/RecOutletWarehouse/RecOutletWarehouse/Models/ITEM_SUBCATEGORY.cs
@@ -10,7 +10,7 @@ namespace RecOutletWarehouse.Models
             this.ITEMs = new List<ITEM>();
         }
 
-        public byte SubcategoryID { get; set; }
+        public short SubcategoryID { get; set; }
         public string SubcategoryName { get; set; }
         public virtual ICollection<ITEM> ITEMs { get; set; }
     }
